Renames session cookie constant for clarity
